@@ -120,23 +120,23 @@
     "@rollup/plugin-sucrase": "^5.0.1",
     "@rollup/plugin-typescript": "^11.1.0",
     "@rollup/plugin-virtual": "^3.0.1",
-    "@sveltejs/eslint-config": "github:sveltejs/eslint-config#v5.8.0",
+    "@sveltejs/eslint-config": "github:sveltejs/eslint-config",
     "@types/aria-query": "^5.0.1",
     "@types/estree": "^1.0.0",
     "@types/mocha": "^10.0.1",
     "@types/node": "^14.14.31",
     "@typescript-eslint/eslint-plugin": "^5.58.0",
-    "@typescript-eslint/parser": "^5.58.0",
     "acorn": "^8.8.2",
     "agadoo": "^3.0.0",
     "aria-query": "^5.1.3",
     "axobject-query": "^3.1.1",
     "code-red": "^1.0.0",
     "css-tree": "^2.3.1",
-    "eslint": "^8.35.0",
+    "eslint": "^8.40.0",
     "eslint-config-prettier": "^8.8.0",
     "eslint-plugin-import": "^2.27.5",
-    "eslint-plugin-svelte3": "^4.0.0",
+    "eslint-plugin-svelte": "^2.28.0",
+    "eslint-plugin-unicorn": "^47.0.0",
     "estree-walker": "^3.0.3",
     "happy-dom": "^9.18.3",
     "is-reference": "^3.0.1",
