@@ -1,6 +1,6 @@
 module.exports = {
 	root: true,
-	extends: ['@sveltejs', 'prettier'],
+	extends: ['@sveltejs'],
 	settings: {
 		'import/core-modules': [
 			'svelte',
@@ -8,8 +8,7 @@ module.exports = {
 			'svelte/store',
 			'svelte/easing',
 			'estree'
-		],
-		'svelte3/compiler': require('./compiler')
+		]
 	},
 	rules: {
 		'@typescript-eslint/no-non-null-assertion': 'off'
