@@ -1,10 +1,10 @@
-import * as fs from 'fs';
-import * as path from 'path';
+import * as fs from 'node:fs';
+import * as path from 'node:path';
 import glob from 'tiny-glob/sync';
 import colors from 'kleur';
 import { assert } from 'vitest';
 import { compile } from '../compiler.js';
-import { fileURLToPath } from 'url';
+import { fileURLToPath } from 'node:url';
 
 export function try_load_json(file) {
 	try {
