@@ -1,8 +1,8 @@
 // @vitest-environment jsdom
 // TODO: https://github.com/capricorn86/happy-dom/issues/916
 
-import * as fs from 'fs';
-import * as path from 'path';
+import * as fs from 'node:fs';
+import * as path from 'node:path';
 import { assert, describe, it } from 'vitest';
 import { create_loader, should_update_expected, try_load_config } from '../helpers.js';
 import { assert_html_equal } from '../html_equal.js';
