@@ -1,6 +1,6 @@
 import { chromium } from '@playwright/test';
-import * as fs from 'fs';
-import * as path from 'path';
+import * as fs from 'node:fs';
+import * as path from 'node:path';
 import { rollup } from 'rollup';
 import { pretty_print_browser_assertion, try_load_config } from '../helpers.js';
 import * as svelte from '../../src/compiler/index.js';
@@ -64,11 +64,11 @@ describe(
 								}
 
 								if (importee === 'main') {
-									return 'main';
+									return '\0virtual:main';
 								}
 
-								if (importee === 'assert') {
-									return 'assert';
+								if (importee === 'assert.js') {
+									return '\0virtual:assert';
 								}
 
 								if (importee === '__MAIN_DOT_SVELTE__') {
@@ -80,9 +80,9 @@ describe(
 								}
 							},
 							load(id) {
-								if (id === 'assert') return browser_assert;
+								if (id === '\0virtual:assert') return browser_assert;
 
-								if (id === 'main') {
+								if (id === '\0virtual:main') {
 									return main.replace('__HYDRATE__', hydrate ? 'true' : 'false');
 								}
 								return null;
@@ -147,7 +147,7 @@ describe(
 					});
 					await page.setContent('<main></main>');
 					await page.evaluate(generated_bundle.output[0].code);
-					const test_result = await page.evaluate(`test(document.querySelector('main'))`);
+					const test_result = await page.evaluate("test(document.querySelector('main'))");
 
 					if (test_result) console.log(test_result);
 					assertWarnings();
