@@ -1,13 +1,13 @@
 // @vitest-environment jsdom
 
-import * as fs from 'fs';
-import * as path from 'path';
+import * as fs from 'node:fs';
+import * as path from 'node:path';
 import glob from 'tiny-glob/sync.js';
 import { beforeAll, afterAll, describe, it, assert } from 'vitest';
 import { compile } from '../../src/compiler/index.js';
 import { clear_loops, flush, set_now, set_raf } from 'svelte/internal';
 import { show_output, try_load_config, mkdirp, create_loader } from '../helpers.js';
-import { setTimeout } from 'timers/promises';
+import { setTimeout } from 'node:timers/promises';
 import { setup_html_equal } from '../html_equal.js';
 
 let unhandled_rejection = false;
@@ -125,7 +125,7 @@ describe('runtime', async () => {
 					});
 
 					try {
-						mod = await load(`./main.svelte`);
+						mod = await load('./main.svelte');
 						SvelteComponent = mod.default;
 					} catch (err) {
 						show_output(cwd, compileOptions); // eslint-disable-line no-console
@@ -148,7 +148,7 @@ describe('runtime', async () => {
 
 						// ssr into target
 						if (config.before_test) config.before_test();
-						const SsrSvelteComponent = (await load_ssr(`./main.svelte`)).default;
+						const SsrSvelteComponent = (await load_ssr('./main.svelte')).default;
 						const { html } = SsrSvelteComponent.render(config.props);
 						target.innerHTML = html;
 
