@@ -26,7 +26,7 @@ export default {
 		deferred = create_deferred();
 		component.thePromise = deferred.promise;
 
-		assert.htmlEqual(target.innerHTML, `<br /><p>the promise is pending</p>`);
+		assert.htmlEqual(target.innerHTML, '<br /><p>the promise is pending</p>');
 
 		const rejection = deferred.promise
 			.catch(() => {})
