@@ -36,13 +36,13 @@ export default {
 		component.visible = true;
 
 		assert.equal(subscribers.length, 1);
-		assert.htmlEqual(target.innerHTML, `<p>value: initial</p>`);
+		assert.htmlEqual(target.innerHTML, '<p>value: initial</p>');
 		value = 42;
 		subscribers.forEach((fn) => {
 			fn(value);
 		});
 		await tick();
-		assert.htmlEqual(target.innerHTML, `<p>value: 42</p>`);
+		assert.htmlEqual(target.innerHTML, '<p>value: 42</p>');
 
 		component.visible = false;
 
