@@ -1,8 +1,8 @@
 // @vitest-environment jsdom
 
-import * as fs from 'fs';
-import * as path from 'path';
-import { setTimeout } from 'timers/promises';
+import * as fs from 'node:fs';
+import * as path from 'node:path';
+import { setTimeout } from 'node:timers/promises';
 import glob from 'tiny-glob/sync';
 import { assert, describe, it } from 'vitest';
 import { compile } from '../../compiler.mjs';
