@@ -1,4 +1,4 @@
-import * as fs from 'fs';
+import * as fs from 'node:fs';
 import { assert, describe, it } from 'vitest';
 import { compile } from '../../src/compiler/index.js';
 import { try_load_json } from '../helpers.js';
